backends: SDL: Avoid lockup in Wait.
Same story again. We cannot sleep in Wait while holding the mutex and we
do not have a file descriptor to poll on so we need resort to polling.

We will need a better solution here.

One posibility is to add a init flag to all backends and resort to
polling only if application needs thread safe backend.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>