loaders: JPG: Don't convert RGB to BGR by default.
The loader was swapping the RGB and BGR in order
return context in standard format.

Now this is not needed as the library handles both RGB and BGR well.

And it also slows things down when the conversion is unnecessary, i.e.
Load image, Resize, Save.

Or when the conversion could be done on resized image, i.e. Load image,
Resize, Blit on the screen (this was about 15% slower on photos from
digital camera).

In order to get this work loaders must be converted to automatic
conversions between RGB888 and BGR888. I was trying to avoid doing any
conversions in loaders but I'm fine with this particular one.

The JPEG loader is already modified to handle both. More to come.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>