filters: Cleanup Resize filters + docs.
This is a major cleanup in resize filters implementation.

The only visible change is that the GP_FilterResize() is split into
allocating and non-allocating filter (as rest of the API was quite a
time ago). The rest of the changes mostly shuffles code and headers.

This commit also fixes the python wrappings for resize filters.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>