filters: Blur: Hotfix the GaussianBlurAlloc() case
The code wrongly passed src as src and dst as dst to both vertical and
horizontal convolution, which is wrong in case src != dst because the
result from the first convolution is always lost.

Workaround this by passing dst as both src and dst to the second one.
The drawback is that the second convolution would not run in multiple
threads but at least it yields correct result.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>