backends: X11: Fix window bitmap buffer init
The w and h may be modified upon the window creation so we must use
wreq.w and wreq.h instead of the user requested w and h.

This fixes bug for X11 backend when root window is used and the image
size was wrongly initialized to whatever user supplied instead of the
root window size.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>