loaders: zip: Partial support for stream zip
Some zip files does not have crc, compressed size and uncompressed size
in zip local header but in data descriptor (which is located after the
data).

This commit modifies the decompression rutines to cope with unknown
buffer size (the buffer is reallocated on the fly).

Seeking in the ZIP does not work yet.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>