spiv: Fix slow blit for images with alpha channel
The pattern_fill() function that draws chessboard like pattern before
images with alpha channel are blitted was drawing the pattern pixel by
pixel, which is quite slow and caused dropped events when moving such
image in the spiv window.

This commit optimizes the function a bit (about then times faster) which
fixes the problem.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>