build: Link gfxprim libs with third party libs
Pass LDLIBS for third party libraries to linker when gfxprim-foo
libraries are created.

This allows us to link binaries only agains the gfxprim-foo libs
(because the third party libs are loaded as dependencies of the
particual gfxprim-foo library) and we don't have to recompile them
when list of libraries GFXprim is linked agains changes.

This commit resolves #6

Signed-off-by: Cyril Hrubis <metan@ucw.cz>