backend: Do not clobber timer->priv on timer event
The backend code that pushed event into the queue on timer expiration
used internally the timer->priv pointer. Which is both confusing and
unfortunate for the user of the interface.

This commit adds _priv pointer to the timer structure to be used by the
backend as well as changes the timer->id to be a pointer instead of an
array.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>