configure: Fix freetype detection
The original header_exists() function was really meant to return true if
at least one of the headers existed. So this reverts the original
function back, which fixes freetype detection.

It also adds headers_exists(), note the plural, which returns true only
if all headers exists, which is then used for the xcb detection.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>