text: Add GP_TEXT_BEARING flag
Sometimes we need to apply bearing before the first letter, sometimes we
do not.

Generally if we write a whole string at once, the bearing should be
ignored, so that the first pixel starts at the passed x position.

However when we want to write a string letter by a letter, the bearing
should be applied, otherwise the letters would be misaligned.

So this comit adds a flag, that if passed, enables bearing before first
letter of the string.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>