widgets: Add proper widget shrink attribute
When widget shrink is disabled widget will be resized only if the new
minimal size is bigger than the old one.

The default value for shrink is 0 for markup and label and 1 for
everything else.

Shrink can be also overriden from JSON by setting the shrink attribute.

As a side effect this commit also optimizes resize handling, if leaf
widget requests resize and it's min_w and min_h are the same the resize
event is ignored now.

TODO: Do we need separate shrink bits for both vertical and horizontal
      direction?

Signed-off-by: Cyril Hrubis <metan@ucw.cz>