Redesing the event queue API
There is absolutely no reason to copy the events on gp_event_queue_get()
and gp_event_queue_peek(), this changes the functions to return a
pointer to an offset in the queue.

Also the global key state shouldn't be part of the event at all, it's
the property of the queue.

+ tests
+ docs

Signed-off-by: Cyril Hrubis <metan@ucw.cz>