widgets: Prevent UB in send event due to type promotion
Clang reports that passing an argument subject to default type
promotion to va_start results in undefined behavior. Because sub_type
is uint16_t and the function is variadic, it will be promoted to
unsigned int.

This can be avoided by simply declaring the parameter as a full width
integer.

Signed-off-by: Richard Palethorpe <richiejp@f-m.fm>