tests: filters: filters_compare: Disable threads
The libc pthread code apparently keeps a few TLS allocated for reuse
later. I guess that this is an optimization that speeds up case when
application is spanwing threads rapidly.

The problem is that it seems that the decision which TLS buffers it
keeps is random, so creating a few threads before we start malloc
tracking does not really help.

So the easiest solution is to disable pthreads at runtime which prevents
false positives but we still keep eye on temporary buffers allocated by
the filters.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>