widgets: tabs: Make the API consistent
It was not clear which operations add/remove widgets and tabs so now all
tabs operation are prefixed with _tab.

E.g. gp_widget_tabs_rem() that removed tab is now renamed to
gp_widget_tabs_tab_rem() and new function has been added to remove a
widget with the name gp_widget_tabs_rem().

This also adds now unused flags parameter to the constructor as I would
like to add support for vertical tabs later on and this avoids breaking
the ABI (since we check that flags has been passed as 0).

+ docs
+ tests

Signed-off-by: Cyril Hrubis <metan@ucw.cz>