widgets: render: Resize window back after dialog
If dialog was bigger than the previous window resize it back to it's
original size.

Also fix some corner cases where we abort() after two dialogs executed
in a row

This basically deffers the backend resize to the redraw function so that
we do not resize the backend when not needed.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>