backends/SDL: Add SDL2 support
The SDL gfxprim backend now defaults to SDL2 if available, i.e. the
backend is compiled and linked against SDL2. If SDL2 is not available or
SDL1.2 was choosen by passing --LIBSDL_VERSION=1 to the configure the
SDL backend is compiled and linked against SDL1.2. This is implement by
a few ifdefs in the backend itself. Note that linking against both at
the same time is not possible since the API is nearly the same and the
function names would clash.

The code to initialize pixmap from surface was moved to a header as a
static inline function and can be used both from SDL and SDL2 at the
same time. The SDL_glue and SDL2_glue demos shows how to do that.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>