utils: htable: Fix remove (again)
The previous code was sublty broken as well, we would have to, in order
to make it work, remove the whole cluster, and then reinsert it one by
one. This would require additional memory and wouldn't be as elegant as
a proper solution.

It turns out this is not needed at all, we have to instead:

* Loop over the rest of the cluster starting with a record right after
  the one we have deleted

* If the hash value is smaller or equal to the removed index, the value
  has to be moved to fill in the empty slot, this creates an empty slot
  and we have to repeat the search for the newly introduced empty slot
  as well

This does work but there is a corner case, we have to consider the case
where the cluster spans across the boundary as the array is usead as a
loop during during the insertion, that means that the indexes overflow
to 0 when we look for an empty slot on insertion. This is for instance
explained in "Sorting and Searching" form KNUTH. I short: we have to
keep the record in place if it's circulary between the empty slot and
the current position in the cluster. The key here is the 'circulary
between' which yields two different cases one when the empty slot index
is smaller than search index and the other one where the search index
overflows and is smaller than empty slot index.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>