text: Add font families
This add font families that group fonts of the same family and size
together. I.e. if we have monospace, regular and bold style for the same
font it goes to the same family.

This also adds a lookup and interators so that we can dynamically
resolve fonts with a proper API instead of hardcoding the pointers. This
would allow us to add compile time options to enable/disable compiled in
fonts.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>