widgets: Fix possible clipboard request race
A malicious application could possibly delay the clipboard request until
the widget has been destroyed which would lead to a crash.

So the tbox widget now attempts to cancel any ongoing clipboard request
which means that we will never attempt to send an event to a destroyed
widget.

I guess that we should also setup a timeout timer for the request as
well, since if we do not get the data in a tenth of a second or so it's
better if it does not arrive at all.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>