text: Rework text width functions
We have two different with funcitons now, one that returns bounding box
and second that returns advance up to a given point in string.

This also fixes the textbox selection "jumping" around since we have a
function to get the exact length to a start of the next chanracter now.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>