widgets: json: Allow to pass callbacks in a struct
This commit adds another parameter to the widget layout JSON loaders
that allows for the on_event callbacks and callback ops to be passed in
an array of structures.

The dialogs are also converted to use this API, which actually
simplifies the code a bit and at the same time removes the -ldl
dependency from them, which makes the dynamic linker optional for the
whole library.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>