Introduce basic support for OS ports
The configure now has os parameter which defaults to linux, based on the
OS parameter different directory in libs/ is used and linked with the
libgfxprim core library. This allows us to link different
implementations for different operating systems.

This commit also moves gp_time_stamp.c to the linux/ directory so that
we can define different implmenetation for different OS.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>