events: Get rid of timeval
This gets rid of struct timeval and gettimeofday() calls and replaces
them with a 64bit integer and gp_time_stamp() calls.

The end result is that comparing two event timestamps is simpler and the
code is much more portable, now only gp_time_stamp() has to be
reimplemented for a given platform.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>