input: eq_queue: Fix gp_ev_queue_push_rel_to()
We need to track two different cursor positions, one that is valid for
the first event in the event queue and one that is valid for the last
event in the queue. Otherwise if we call gp_ev_queue_push_rel_to() more
than once we will use old position valid for the first event in the
queue and the cursor will jump more than it should.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>