backends: display: uc8179: Add LUTs for faster mode
- full refresh now takes around 3 seconds
- we have very fast partial refresh

The downside is that this produces a bit more noise than doing big full
refresh all the time, but this mode is quite nice for an interactive
mode e.g. terminal.

Also I'm not sure if it will still work with temperatures that are
significatnly lower than the room temperature.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>