widgets: table: Fix click on keyup
Due to changes that implemented scrolling by dragging we implemented
click that selects a table row on a key up rather than on key down.
However we have to check if there was a key down event for the table
otherwise we will do a bogus row selection under certain circumstances.

I.e. if dialog appears because we clicked on a button the keydown event
will be caught by the dialog, which in turn fixes crashes with file open
dialog where we attempt to get filename from dir cache before it was
initialized.

Fixes: eb155e823da1 ("widgets: table: Add scrolling by dragging")
Signed-off-by: Cyril Hrubis <metan@ucw.cz>