backends/display: Change GPIO pwr mapping
Waveshare e-ink displays does not use pwr GPIO18 at all, let's remap
this to GPIO13 that has hardware PWM so that we can use it later to
drive leds that will shine down on the display.

This also allows us to drive the BLK pin of st77XX displays with
hardware PWM so that we can regulate backlight brightness.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>