core: get_put_pixel: Fix PixelPack endianity
The problem is that pixels such as RGB565 has to have bytes stored in
the buffer in the order of increasing array indexes i.e. Big Endian. But
for pixel types which have a channels spanning two bytes e.g. 16BPP
Grayscale we have to store pixels in the machine endianity so that we
can fetch them from memory in the right order.

All of this means that we have to explicitly define endianity for
PixelPacks that are spanning more than one byte and that there are three
types of endianity to be set, BE, LE and ME (machine endian).

Signed-off-by: Cyril Hrubis <metan@ucw.cz>