backends: proxy: Fix dropped update rects
If the client fills in the queue with update rect requests the sendmsg()
starts to fail because the file descriptor is opened O_NONBLOCK and we
even pass MSG_DONTWAIT.

This fixes the problem by making the client to block if the server queue
is full.

At the same time we set the client fd to O_NONBLOCK so that when we
write input events to a client a misbehaving client cannot lock the
proxy backend.

TODO: Force quit an client when it fails to respond to a quit request.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>