loaders: zip: Search for a start of the first record
There seems to be a zip files in the wild that have random mess between
the zip file hader "PK" and start of the first record "PK\x03\x04".
Other zip implementation seems to work with these just fine, since they
use the central archive directory at the end of the file to get all the
offsets. We parse the file from start in order to be able to be work
with incomplete and damaged files. In order to be able to parse these
files we now look for the first record header and ignore bytes before
before it.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>