loaders: exif: Fix byte and short load with MM
The value if it fits into 4 bytes is always aligned to the left. This
means that if the Exif is stored in big endian (MM) we get the bytes in
wrong alignment but right order.

For 1 byte stored in 4 bytes we get:

             Machine Endian  Exif Endian  Cast OK
BB xx xx xx       II            II         Yes
BB xx xx xx       MM            MM         No
xx xx xx BB       II            MM         No
xx xx xx BB       MM            II         Yes

That is because if Machine endian does not equal Exif endian we swap the
4 bytes. For the cast to work the byte has to be on the left for Machine
Little endian (II) and on right for Machine Big endian.

And similarily for the short 2 bytes case.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>