backends: proxy: Make gp_backend_update_rect() MT safe
This commit fixes update_rect() so that we can have up to 16 threads
calling gp_backend_update_rect() simultaneously and all the threads
would be properly blocked until the backend finishes copying the data.

This is done by:

Adding mutex around the function that reads the the messages from
proxy backend. With that only one thread is reading and parsing the
messages from the server.

Adds cookie for each GP_PROXY_UPDATE message that is echoed back by the
proxy backend to the application, this allows the the application
properly pair the completion notification with particular request. The
cookie is just number that increases for each gp_backend_update_rect()
request.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>