backends: linux touchpad: Add filtering + scaling
This fixes touchpad noise and slows it down so that touchpad with Linux
DRM backend is actually usable.

First we apply adaptive low pass filtering. The idea is simple the
closers the samples the heavier filtering can be applied since with
closer samples the mouse has smaller velocity. This smooths down the
mouse pointer while keeping it responsive to larger jumps. In the end
the low pass filter is only applied when samples are closer than a
certain threshold, which removes jitter completely.

At the same time the mouse moved too fast, which made it harder to use,
so we simply scale down the result.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>