backends: Properly define flip and update semantics
This has been poorly defined from the start and needed to be clarified.

Now the sematics is:

- gp_backend_flip() swaps bugger between display buffer and backend
  buffers

- gp_backend_update*() copies data to a display

This means that unlike gp_backend_flip() the backend pixmap buffer data
are kept after the gp_backend_update*().

And of course depending on which kind of display we draw to flip may
fall back to update and the other way around. Or may even be no-op.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>