libs/input: Make the UTF event part of a key event
The keypresses and unicode characters should be delivered in a single
event. So now the key event has two components now. One is a key
scancode and second is a unicode character.

Syntetic unicode characters have the key set to 0 and keypresses that
does not generate unicode characters have utf set to 0.

If a key press generates more than one unicode characters the last
unicode character has the key set to the scancode key.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>