backends: redesing resize_ack()
Make the rendering event based. Previously after backend has been
initialzied the backend->pixmap was used to determine pixel type.

Now there is a GP_EV_SYS_RENDER_PIXEL_TYPE vent that is send to
applicaton once pixel type has been determined so that application
can prepare the colors.

Then GP_EV_SYS_RENDER_RESIZE event is send and the application can
prepare the rendering for resize.

And the rendering is finally started by GP_EV_SYS_RENDER_START from that
point on the backend->pixmap pointer is valid and can be used for
rendering.

When the backend pixmap needs to be resized or when the application is
not shown on the screen the GP_EV_SYS_RENDER_STOP event is send, which
is acked by the application with gp_backend_render_stopped(). The
backend->pixmap pointer is set to NULL during the time rendering is
stopped.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>